Revert "KFSPTS-8339: Extracted common features to helper superclass."
This reverts commit 2e44fd3064d96b756b426dc07a81675a4a78b9a2.
